sched: idle: cpuidle: Add debugfs to check prediction correctness
Evaluating the gap between the bet for the next event waking up a cpu and
the effective sleep duration is not possible today.

This simple patch provides a set of files to evaluate if the predictions
are correct or not.

There are splitted in 3 categories:

 * under estimated : the sleep duration was greater than what we were expecting,
   so we could have been deeper in the idle state. That is true, if the deeper
   idle state was not disabled and fitting the exit latency requirement.

 * over estimated : the sleep duration was smaller than what we were expecting and
   we did not reach the break even for the selected idle state.

 * well estimated : the excepted sleep duration lead to the correct idle state
   selection and we got an optimum power energy saving.

These informations are exported in the debugfs in the sched/idle directory.

Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>