arch/arm: Add qemupalmetto dts with Cadence GEM support
Describes a machine that is as similar to Palmetto, primarily for running under
QEMU. Relaxing the constraint of booting the exact Palmetto hardware in QEMU
allows us to integrate existing device models to expand the machine's
capability without huge effort.

In this case, the Cadence GEM controller is integrated as there was an existing
model in QEMU which was easy to integrate (both in QEMU and kernel).

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>