fsi/master-gpio: implement finer-grained locking over command transactions
Currently, the GPIO master driver grabs cmd_lock at the start of a FSI
transaction, and holds it until the entire transaction is complete. This
may mean that we hold it over (up to 100) DPOLL retries, if the slave
device reports busy.

This change modifies the locking semantics a little, by only holding the
spinlock over clocking the bits in/out of the bus. This way, we drop the
lock once we've sent a command, and reacquire it when we want to read a
response.

We need to ensure that a request write is followed by the corresponding
response read. To prevent interleaved commands, we introduce a mutex
over the whole ->read / ->write / ->term operation.

WIP: testing needed, and we may want to move the cmd_lock mutex up into
the FSI core, as it's likely to be needed by other master
implementations. Also, we may be able to replace the spinlock with
local_irq_*.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>