pinctrl: aspeed-g6: Make the SIG_DESC_CLEAR() behave intuitively
Signal descriptors can represent multi-bit bitfields and so have
explicit "enable" and "disable" states. However many descriptor
instances only describe a single bit, and so the SIG_DESC_SET() macro
is provides an abstraction for the single-bit cases: Its expansion
configures the "enable" state to set the bit and "disable" to clear.

SIG_DESC_CLEAR() was introduced to provide a similar single-bit abstraction for
for descriptors to clear the bit of interest. However its behaviour was
defined as the literal inverse of SIG_DESC_SET() - the impact is the bit
of interest is set in the disable path. This behaviour isn't intuitive
and doesn't align with how we want to use the macro in practice, so make
it clear the bit for both the enable and disable paths.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>