mctp: Add MCTP-over-KCS transport binding
This change adds a MCTP KCS transport binding, as defined by the DMTF
specification DSP0254 - "MCTP KCS Transport Binding". An MCTP protocol
network device is created for each KCS channel found in the system. The
interrupt code for the KCS state machine is based on the current IPMI
KCS driver.

Signed-off-by: Konstantin Aladyshev <aladyshev22@gmail.com>